feat: statistics

+++ src/main/java/com/cloudians/domain/personaldiary/service/PersonalDiaryService.java
@@ -24,13 +24,12 @@ import com.cloudians.global.exception.JsonExceptionType;
 import com.cloudians.global.service.FirebaseService;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.firebase.auth.FirebaseAuthException;
+import com.google.cloud.storage.Blob;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.IOException;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.Arrays;
@@ -88,7 +87,7 @@ public class PersonalDiaryService {
         return PersonalDiaryEmotionUpdateResponse.of(editedEmotions);
     }
 
-    public PersonalDiaryCreateResponse createPersonalDiary(PersonalDiaryCreateRequest request, String userEmail, MultipartFile file) throws Exception {
+    public PersonalDiaryCreateResponse createPersonalDiary(PersonalDiaryCreateRequest request, String userEmail, MultipartFile file) {
         User user = findUserByUserEmail(userEmail);
         //감정 수치 가져오기
         PersonalDiaryEmotion emotions = getTempEmotion(user.getUserEmail());
@@ -116,7 +115,7 @@ public class PersonalDiaryService {
         return PersonalDiaryResponse.of(personalDiary);
     }
 
-    public PersonalDiaryResponse editPersonalDiary(PersonalDiaryUpdateRequest request, Long personalDiaryId, String userEmail, MultipartFile file) throws Exception {
+    public PersonalDiaryResponse editPersonalDiary(PersonalDiaryUpdateRequest request, Long personalDiaryId, String userEmail, MultipartFile file) {
         User user = findUserByUserEmail(userEmail);
         // 수정할 일기가 있는지 확인
         PersonalDiary personalDiary = getPersonalDiaryOrThrow(personalDiaryId, user);
@@ -131,11 +130,13 @@ public class PersonalDiaryService {
         return PersonalDiaryResponse.of(editedPersonalDiary);
     }
 
-    private String updateDiaryPhoto(String userEmail, MultipartFile file, PersonalDiary personalDiary) throws Exception {
+    private String updateDiaryPhoto(String userEmail, MultipartFile file, PersonalDiary personalDiary) {
         String fileName = getFileName(personalDiary.getDate());
         try {
-            firebaseService.getFileUrl(userEmail, DOMAIN, fileName);
-            firebaseService.deleteFileUrl(userEmail, DOMAIN, fileName);
+            Blob existingBlob = firebaseService.getBlob(userEmail, DOMAIN, fileName);
+            if (existingBlob != null && existingBlob.exists()) {
+                firebaseService.deleteFileUrl(userEmail, DOMAIN, fileName);
+            }
         } catch (FirebaseException ignored) {
         }
         return uploadPhoto(userEmail, file, personalDiary.getDate());
@@ -175,6 +176,7 @@ public class PersonalDiaryService {
         User user = findUserByUserEmail(userEmail);
         PersonalDiary personalDiary = getPersonalDiaryOrThrow(personalDiaryId, user);
 
+        validateIfPhotoExistsOrThrow(personalDiary);
         deletePhotoIfExists(userEmail, personalDiary);
         personalDiary.deletePhotoUrl();
     }
@@ -211,13 +213,13 @@ public class PersonalDiaryService {
     }
 
     private String extractValue(String[] lines, int index) {
-        if (lines.length <= index || !lines[index].contains(ANSWER_START_REGEX)) {
+        if (lines.length <= index) {
             throw new PersonalDiaryException(PersonalDiaryExceptionType.WRONG_CHATGPT_ANSWER_FORMAT);
         }
         return lines[index].split(ANSWER_START_REGEX)[1].trim();
     }
 
-    private List<String> getElementCharacters(FiveElement element) {
+    public List<String> getElementCharacters(FiveElement element) {
         return fiveElementCharacterRepository.findRandomCharactersByElementId(element.getId()).stream().map(FiveElementCharacter::getCharacter).collect(Collectors.toList());
     }
 
@@ -234,7 +236,6 @@ public class PersonalDiaryService {
     }
 
     private void deletePhotoIfExists(String userEmail, PersonalDiary personalDiary) {
-        validateIfPhotoExistsOrThrow(personalDiary);
         if (personalDiary.getPhotoUrl() != null) {
             firebaseService.deleteFileUrl(userEmail, DOMAIN, getFileName(personalDiary.getDate()));
         }
@@ -246,8 +247,8 @@ public class PersonalDiaryService {
         }
     }
 
-    private String uploadPhoto(String userEmail, MultipartFile file, LocalDate diaryDate) throws IOException, FirebaseAuthException {
-        return file != null ? firebaseService.uploadFile(file, userEmail, getFileName(diaryDate), DOMAIN) : null;
+    private String uploadPhoto(String userEmail, MultipartFile file, LocalDate diaryDate) {
+        return file != null ? firebaseService.uploadFile(file, userEmail, DOMAIN, getFileName(diaryDate)) : null;
     }
 
     private String getFileName(LocalDate diaryDate) {
